Use the same computation to compute geothermal renewables for germany and municipalities.
